+++ page_switcher/python/Windows_SysTray/image.ico
