+++ page_switcher/python/auto_page_switch.py
